Weekly Calendar PR fix
